Change public interface such that WebGameCreatep#play will call #play_move
WebApp was calliing play then play_move which seemed like it needed to
know how to build a WebGame rather than leaving it up to #play.

It's a tradeoff as now WebGameCreate#play now has a conditional which
only is true when 'params' has a 'position'.
I'm thinking that since WebGameCreate has the verb in it's name that
maybe the TicTacToe::Game instance could be created in the WEbGameCreate
initialize constructor.

What do you think?  @maikon and @jsuchy